+++ packages/api/cms-api/src/common/decorators/crud-generator.decorator.ts
@@ -2,12 +2,13 @@ import { type Type } from "@nestjs/common";
 
 import { type CurrentUser } from "../../user-permissions/dto/current-user";
 import { type Permission } from "../../user-permissions/user-permissions.types";
+import { type MutationError } from "../graphql/mutation-error";
 
 export interface CrudGeneratorHooksService {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    validateCreateInput?: (input: any, options: { currentUser: CurrentUser; scope: any; args: any }) => Promise<void>;
+    validateCreateInput?: (input: any, options: { currentUser: CurrentUser; scope: any; args: any }) => Promise<MutationError[]>;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    validateUpdateInput?: (input: any, options: { currentUser: CurrentUser; scope: any; entity: any }) => Promise<void>;
+    validateUpdateInput?: (input: any, options: { currentUser: CurrentUser; scope: any; entity: any }) => Promise<MutationError[]>;
 }
 
 export interface CrudGeneratorOptions {